apps/processor/pyproject.toml
[tool.poetry]
name = "processor"
version = "1.0.0"
description = "Neuro Data Platform - Raw Sensor Data Processor Service"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
psycopg = { extras = ["binary"], version = "^3.1.19" }
minio = "^7.2.7"
zstandard = "^0.22.0"
numpy = "^1.26.4"
python-dotenv = "^1.0.1"

[tool.ruff]
# Inherit all settings from the root pyproject.toml
# This file can be used to override settings for this specific service if needed.
